Refactored requests reducers into more performant queries reducer
