+++ packages/redux-saga-requests/src/reducers/network-reducer.js
@@ -4,7 +4,7 @@ import {
   getRequestActionFromResponse,
 } from '../actions';
 import defaultConfig from './default-config';
-import requestsReducer from './requests-reducer';
+import queriesReducer from './queries-reducer';
 import mutationsReducer from './mutations-reducer';
 import cacheReducer from './cache-reducer';
 
@@ -12,19 +12,7 @@ export default localConfig => {
   const config = { ...defaultConfig, ...localConfig };
 
   return (state = { queries: {}, mutations: {}, cache: {} }, action) => {
-    const queries = Object.entries(state.queries).reduce(
-      (prev, [actionType, query]) => {
-        prev[actionType] = requestsReducer(query, action, actionType);
-        return prev;
-      },
-      isRequestAction(action) &&
-        config.isRequestActionQuery(action) &&
-        !(action.type in Object.keys(state.queries))
-        ? {
-            [action.type]: requestsReducer(undefined, action, action.type),
-          }
-        : {},
-    );
+    const queries = queriesReducer(state.queries, action, config);
 
     let { mutations } = state;
